+++ VerticalToolbar/manifest.json
@@ -1,0 +1,13 @@
+{
+  "Author": "Brandon Marquis Markail Green (Space Baby)",
+  "Description": "Enables an extra vertical bar",
+  "EntryDll": "SB_VerticalToolMenu.dll",
+  "Name": "Vertical Toolbar",
+  "UniqueID": "SB_VerticalToolMenu",
+  "Version": {
+    "MajorVersion": 1,
+    "MinorVersion": 3,
+    "PatchVersion": 4,
+    "Build": ""
+  }
+}
